fix: resolve erro de requisição para a API do Strapi
