+++ src/app/page.tsx
@@ -1,6 +1,4 @@
-"use client";
 
-import { Tab } from "@headlessui/react";
 import ByronCard from "../components/ByronCard";
 import CallButton from "../components/CallButton";
 import axiosHeader from "@/api/axiosHeader";
@@ -9,10 +7,7 @@ import axiosHeader from "@/api/axiosHeader";
 import { IHomepage, Attributes } from "@/interface/IHomepage";
 import IOrganizers from "@/interface/IOrganizers";
 import ISchedules from "@/interface/ISchedules";
-
-function classNames(...classes: any) {
-  return classes.filter(Boolean).join(" ");
-}
+import Programacao from "@/components/Programacao";
 
 interface IProps {
   homepage: Attributes;
@@ -20,7 +15,7 @@ interface IProps {
   schedules: ISchedules[];
 }
 
-export const getStaticProps = async () => {
+const APIRequests = async () => {
   try {
     // HOMEPAGE
     const homepageRes = await axiosHeader.get<IHomepage>(
@@ -40,15 +35,14 @@ export const getStaticProps = async () => {
     );
     const schedules = schedulesRes.data.data;
 
-    return {
-      props: { homepage, organizers, schedules },
-    };
+    return { homepage, organizers, schedules }
   } catch (error) {
     throw error;
   }
 };
 
-export default function Home({ homepage, organizers, schedules }: IProps) {
+export default async function Home() {
+  const { homepage, organizers, schedules } = await APIRequests()
   return (
     <>
       {/* HERO */}
@@ -130,57 +124,7 @@ export default function Home({ homepage, organizers, schedules }: IProps) {
       {/* /SOBRE */}
 
       {/* PROGRAMAÇÃO */}
-      <section
-        id="programacao"
-        className="flex items-center justify-center bg-gradient-to-b from-blue-dark to-blue-light text-white py-12 pb-36"
-      >
-        <div className="flex flex-col items-center max-w-6xl w-full">
-          <Tab.Group>
-            <div className="flex items-center rounded-3xl bg-gray shadow-xl">
-              <Tab.List as="ul" className="flex flex-col items-start gap-1">
-                {schedules.map((tab, index) => (
-                  <Tab
-                    as="li"
-                    key={index}
-                    className={({ selected }) =>
-                      classNames(
-                        "bg-blue-dark rounded-r-2xl first:rounded-tl-2xl first:rounded-tr-none last:rounded-bl-2xl last:rounded-br-none shadow-lg w-80",
-                        selected
-                          ? "bg-blue-light w-[360px] z-10 first:mt-0 my-1 last:mb-0"
-                          : ""
-                      )
-                    }
-                    // className="bg-blue-dark rounded-r-2xl first:rounded-tl-2xl first:rounded-tr-none last:rounded-bl-2xl last:rounded-br-none shadow"
-                  >
-                    <button className="flex flex-col items-start gap-2 text-xl px-6 py-3 text-left">
-                      <h3 className="text-2xl font-bold">
-                        {tab.data[0].attributes.tabDay}
-                      </h3>
-                      <span>{tab.data[0].attributes.organizer}</span>
-                      <p className=" font-bold">
-                        {tab.data[0].attributes.theme}
-                      </p>
-                    </button>
-                  </Tab>
-                ))}
-              </Tab.List>
-
-              <Tab.Panels>
-                {schedules.map((tab, index) => (
-                  <Tab.Panel
-                    key={index}
-                    className=" bg-blue-light m-16 p-6 rounded-3xl"
-                  >
-                    <h2>{tab.data[0].attributes.organizer}</h2>
-
-                    <p>{tab.data[0].attributes.theme}</p>
-                  </Tab.Panel>
-                ))}
-              </Tab.Panels>
-            </div>
-          </Tab.Group>
-        </div>
-      </section>
+      <Programacao />
       {/* /PROGRAMAÇÃO */}
 
       {/* MAPS */}
